Implement Flask application and build pipelines
This commit introduces a Flask web application in the app.py file which exposes an HTTP endpoint for data conversion. It uses Flask CORS and Prometheus Metrics to handle cross-origin requests and gather metrics respectively. The app is supplemented with a requirements.txt file for easy environment setup.

A Dockerfile was added to create a contained environment for the application and a .gitignore file to exclude specific files and directories from version control system such as IDE settings and bytecode.

Additionally, a comprehensive Jenkins file configures the continuous integration / continuous delivery pipeline from building a Docker image to pushing it to a Docker registry and deploying the updated service to a remote host.

These files together form the base of our development environment and continuous deployment strategy.