Update url_repository from hardcoded to environment variable
This commit changes the hardcoded url_repository in the Flask application to use an environment variable. The previous hardcoded value was 'http://127.0.0.1:5001' but it has been changed to fetch from the environment variable 'URL_REPOSITORY'. This change improves flexibility and conforms to best practices, specifically the twelve-factor app methodology for building scalable and maintainable applications.