Update Jenkinsfile credentials and host information
Altered the Jenkinsfile to use credentials and host details from 'one_click_host' instead of 'developHost'. This change is done to improve the security and adaptability of the code. Previous hardcoding of these details resulting in inflexibility and potential exposure. Now, the new host will be fetched from the environment variable 'ONE_CLICK_HOST_IP', strengthening the code with dynamic data injection and safekeeping sensitive data.