Integrate headers management with correlation id
The handling of HTTP headers within the application has been revised, adding management of correlation id in case this value is available. This allows for better request tracking and more detailed diagnostics in distributed systems. If the correlation id isn't available, the application will proceed with the existing Authorization header.